Removed polyglot part-of-speech separation
Previously, we were using this both for detecting proper nouns (to color red) as well as boundaries to do line breaks. This resulted in issues with punctuation, so for now I've reverted to splitting the sentences up on whitespace instead. Sadly this means no text highlighting for now, but the text should look better otherwise.